Support decryption of too long/cut off lines
Sometimes clients don't split too long lines on their end causing the
encrypted ciphertext to be too long and thus cut off by the ircd.
These lines subsequently fail the base64 decode validation because they
won't be padded correctly.
With these changes lines that have invalid padding will be decrypted
anyway and (optionally) a notification will be added on top of the line
to indicate that the decryption is based on cut off input.