Fix handling for buffers with . when using markers
The previous code made too many assumptions about buffer names.
Keys are stored in a `server/channel` or `server/nick` fashion which is
not generally a problem. But weechat-internal the buffers are stored as
`server.nick` or `server.#channel`.
The previous code used the weechat buffername and replaced . to / to
build the identifier for keys in this plugin. Problem is, if you have a
buffername that contains a . e.g. `my.server` or a channel like
`#.chan` you'd end up with a reference to a key that does not exist
(`server/#/chan`)
The only place this is actually a problem in is in the marker handling.
The code would due to not finding the key reference assume user input
would not need the marker (`fish_modifier_input_text`) and thus not add
it to a outgoing message.
In `fish_modifier_out_privmsg_cb` on the other hand the logic is
different and would correctly determine that the text should be
encrypted and call `fish_msg_wo_marker` which based on the settings
would then remove the marker from the message before encrypting it and
sending it to the server.
Since the marker was never added on client side in the first place it
instead removes parts of the outgoing message based on the marker
length and thus cuts off the outgoing message. For the user the message
is displayed correctly but for other users the message would be cut and
missing as many chars as the marker length would've been.