Fix input handler possibly botching messages
The input handler uses the current buffer to check if a fish key is set
and uses that to decide whether or not the message being sent should
have a marker or not. The logic does not make sense and breaks if
scripting is used to send messages to a buffer that is not in focus.
The in-focus buffer is used to determine if the message to this other
buffer uses fish or not and thus needs a marker or not. If the current
buffer does not use fish but the target buffer uses fish the marker will
_not_ be appended. The logic in `fish_modifier_out_privmsg_cb` does not
know this and removes the marker before sending the message and thus
essentially cutting off parts of the message because the marker was not
added before.